+++ server/src/constants.ts
@@ -52,4 +52,3 @@ export let bsconfigModuleDefault = "commonjs";
 export let bsconfigSuffixDefault = ".js";
 
 export let configurationRequestId = "rescript_configuration_request";
-export let pullConfigurationInterval = 10_000;